Several fixes related to task management and cache
This reverts a bunch of earlier work, but changes the way that task
switching works. Instead of trusting the exiting task to leave the
memory space in a usable state we prepare the memory area on entering
rather than cleaning it up on exiting.

Thanks to Peter we also uncovered a bug where process starting was
racey. It is entirely possible that this caused issues as well. In the
previous setup it was possible for a task to get suspenced during setup,
after which it would want to get mapped/remapped, potentially before the
task_info struct was properly installed.

At least, that is the theory. This fixes a bug whereby a program that
almost immediately exits crashes.