Fix window cleanup
Previously we weren't actually deallocating the framebuffer pages, so
there was a page leak. In addition due to the mmu state checking we got
a crash because if a new framebuffer wants to reuse the same vaddr we
can't.

Additionally fix a double free, and a free on a static buffer.