Occlusion culling for the compositor
We now generate a list of visible regions for each window. This list
gets updated whenever a window appears, disappears, or moves.

Due to a (hardware?) problem of the PPA, if we give it input rects of
size (height > 32 && (height % 32) == 1) the PPA hardware seems to
crash. So we carefully post-process the new rects to ensure that
pre-scaled we never read anything of that size. We only ever read 31 or
34 sized stripes, never 33.

Additionally, we now have a messaging system to the compositor, this
means that low-priority tasks won't have to wait for the window_list
lock to be released in order to get a window.

We have also introduced a guard page in between framebuffers, so if a
user accidentally writes past their framebuffer they will crash and not
corrupt another frame buffer.

And a whole bunch of fun rect maths.